+++ dependency/core/processor/classifier_processor.py
@@ -4,7 +4,7 @@ from .processor import Processor
 
 from core.lib.estimation import Timer
 from core.lib.content import Task
-from core.lib.common import Context, LOGGER, ClassFactory, ClassType
+from core.lib.common import Context, LOGGER, ClassFactory, ClassType, FileOps
 
 
 @ClassFactory.register(ClassType.PROCESSOR, alias='classifier_processor')
@@ -15,7 +15,7 @@ class ClassifierProcessor(Processor):
         self.classifier = Context.get_instance('Classifier')
 
     def __call__(self, task: Task):
-        data_file_path = Context.get_temporary_file_path(task.get_file_path())
+        data_file_path = FileOps.get_task_file_in_temp(task)
         cap = cv2.VideoCapture(data_file_path)
         content = task.get_prev_content()
         if content is None: