+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Bug Fix
 - Fix iptables rule accumulation for edgemesh in incorrect dayu shutdowns with `dayu.sh` script correction.
+- Separate task temporary directory for different users (`controller`/`processor`).
 
 ### Minor Update
 - Update log export mode to support large logs in multi-stream scenarios (`backend`/`distributor`).